hwdrm-videoproc-ta: Add TA for CENCv1 H264 handling
This adds an OP-TEE TA that will be using for auxiliary HWDRM video
processing. The first function is CENCv1 H264 slice header parsing.

This receives a structure on the SPS/PPS data and also a handle to the
secure buffer which has the slice data in it. It then parses the slice
data and populates a returned structure. This is used by Chrome via the
cdm-oemcrypto daemon to handle CENCv1 H264 content in V4L2.

Later a fuzzer and unit tests will be added for the parser.

BUG=b:248610688
TEST=H264 CENCv1 plays back on geralt

Change-Id: Ib8934aa1bfa4dd1a6f1c5afe943100504df7329e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4687883
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>