shill: Add permanent MAC to DeviceInfo
In Linux kernel 5.5 a new attribute with permanent/hardware address has
been added to the link info messages (IFLA_PERM_ADDRESS).  This change
is being backported to all ChromeOS kernel versions so let's use it.

This attribute should always be available (for real devices) but if it
is missing we will fall back to ethtool ioctl (the one used in "ethtool
-P wlan0").

BUG=b:258208400
TEST=FEATURES=test emerge-$BOARD shill
TEST=Manual.  Deploy to the board restart and check that this log:
      LOG(WARNING) << "Perm MAC requested for device w/o it in info: "
    is not present.
    Note: if the board is running on kernel version prior to 5.5 you
    need to update the kernel to the version that includes one of the
    backports listed in:
    https://chromium-review.googlesource.com/q/topic:%22stable-storage-id%22

Cq-Depend: chromium:4775598
Change-Id: I9c1f87c841afd1b373ae2c8ac24d49bc3367d58e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4687458
Reviewed-by: David Ruth <druth@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Andrzej Ostruszka <andrzejo@google.com>