shill: wifi: initiate a scan in self-managed solution when starting ap interface
For self-managed WiFi solution, a scan is needed to fetch the current
region the device is in. Initiate a scan during AP interface enablement
if the WiFi solution is region self-managed.

BUG=b:281130838
TEST=FEATURES=test emerge-${BOARD} shill
TEST=on brya (region self-managed), start tethering session for multiple
times. Tethering can be start/stop successfully. Check log and see if
scan is run and region has been updated after that.

Change-Id: I4afd4d60f779990a6090afe3291d2a2875c83d92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4687236
Tested-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Jintao Lin <jintaolin@chromium.org>