power: Update retry for PowerSupply::RefreshImmediately
Sometimes reading power supply data can be fail and we
have retry code for that with CHECK at the end that it
shouldn't fail with retries. But there are still some
crashes there.

Current retry delays are 1, 2 and 4 ms. But power_supply
sysfs is actually updated every second.

This CL does the following
* Update retry delay to 1, 10, 100, 1000 and 1000 ms
  to ensure that the delays are long enough.
* Add PowerSupply::RefreshImmediatelyWithRetry to use
  the same code in both dump_power_status and
  power_supply_info

BUG=b:287889424
TEST=CQ

Change-Id: If58a3a47100bbf16c1e726133b5bc1adf473a51c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4685137
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Puthikorn Voravootivat <puthik@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Puthikorn Voravootivat <puthik@chromium.org>