chromeos-config: Add block-floss-availability flag
This adds a flag in bluetooth/flags to disable Floss in certain
models.

The reason why we add this is because we want to enable Floss on all
brya boards except for marasov via finch. However, the finch config is
not capable to filter devices by model. Another solution is to check the
model and block it in chrome side, but this will change many places so
this temporary solution is needed.

BUG=b:285808593
TEST=build

Change-Id: I9063feec4b4e16b03b498aef9c953085144897c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4686892
Commit-Queue: Yun-Hao Chung <howardchung@chromium.org>
Tested-by: Yun-Hao Chung <howardchung@chromium.org>
Reviewed-by: Jesse McGuire <jessemcguire@google.com>