cryptohome: Make InvalidateAuthSession non-blocking
This call currently always succeeds, and we're going to update Chrome
such that it doesn't block on the response of InvalidateAuthSession. To
get the precise latency metrics in perf tests, update the CLI call to
simply return after sending the request.

BUG=b:287305183
TEST=tast run cryptohome.PinLockoutOnLockScreen
TEST=tast run hwsec.AuthSessionLightweightAuthPerf, the latency is low
as usual even if there's a PIN to reset.

Change-Id: I6c17996fea329b0c10262a2da84ff7d9300e0a19
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4685837
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>