camera: remove failing DCHECK
We are explicitly marking the result buffers as
CAMERA3_BUFFER_STATUS_ERROR, with the expectation that this will get
cleared upon successful processing through the effects pipeline.

As such, we can't check at the top of RenderEffect that the buffer is
OK. This is, however, already checked at :853 so we don't run the risk
of trying to process a previously bad buffer.

BUG=b:291658845
TEST=run tests on DUT

Change-Id: Ib9d325301b939803bd26955acd933f1b6e63983b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4693104
Auto-Submit: Jim Pollock <jmpollock@chromium.org>
Commit-Queue: Sami Kyöstilä <skyostil@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Tested-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>