dlp: Refactor to define FileId type (inode).
Refactoring to prepare to change using inode as file identifier
in DLP daemon, usages of `ino64_t` type and `inode` name
are changed to neutral FileId and `id` respectively.

As a next step, `FileId` will be changed to contain both
inode and crtime of the file.

BUG=b:290913473
TEST=FEATURES=test emerge-{$BOARD} chromeos-base/dlp

Change-Id: I265912cd57326bf7a5f96799b759f98b0674eece
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4693568
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Tested-by: Sergey Poromov <poromov@chromium.org>