crosh: Inform the user when hibernate is not available and why
Hibernate is currently only supported on certain boards and
even on these boards it needs to be explicitly enabled
(through chrome://flags or finch). Further hibernate requires
a user to be logged in, so it is not supported in guest mode.
Lastly the 'hiberimage' might not exist, which would be the
case if a user previously logged out.

Check for the above conditions and tell the user when hibernate
is not available and why.

BUG=b:291068562
TEST=run 'hibernate' in crosh with different configurations

Change-Id: I0918ed49fcbcdab6a4c86da13f95c553aa8738be
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4706484
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>