dns-proxy: always destroy the ARC Cellular proxy
This patch makes sure that dns-proxy destroy the DNS proxy instance
created to the ARC Cellular Device when that Device is advertised as
destroyed by patchpanel.

This ensures that the ConnectNamespace requests sent by dns-proxy to
patchpanel are revoked when the shill Cellular Device becomes
disconnected, and this avoids the need for patchpanel to dynamically
update a ConnectNamespace bound to a Cellular Device that disconnects
and loses its primary multiplexed interface.

crrev/c/4190917 added logic to avoid destroying the ConnectNamespace of
a dns-proxy instance when an ARC virtual device is destroyed in order to
limit IPv6 address churn (b/266496850).

But for Cellular Devices, b/266496850 is not expected to happen and this
logic can be ignored and it is not necessary to avoid destroying the
ConnectNamespace and maintain the previous IPv6 address used by
ConnectNamespace.

BUG=b:266496850
BUG=b:273741099
TEST=FEATURES=test emerge-$BOARD dns-proxy

Change-Id: I8625ae4bb2a11cf67f323f0c284e4279f4549de7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4699524
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>