Add extension.log file for system extension API logs
Redirect the logs added by extensions using the the system log API
from /var/log/messages to a separate log file
(/var/log/extensions.log). Also add the file to feedback reports.
The API is limited to system extensions developed by Google, and is
currently limited to one extension (privacy approval for making its
logs persistent: b/279975328).

The new file is limited to the customers using the extension that are
going through an initial deployment and debug phase (the file will
improve the debugging experience of our support and CEs). Logs will
not be added any more after the debugging phase is finalized and the
extension is deployed to production.

BUG=b:291047138
TEST=filed feedback and confirmed that log file is present, also
checked my device for the file and if the logs are correctly
redirected to it

Disallow-Recycled-Builds: test-failures
Change-Id: I75057a9e18902777668a51ddf62e5f3e3faf5837
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4701643
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Yoshiki Iguchi <yoshiki@chromium.org>
Commit-Queue: Maria Petrisor <mpetrisor@chromium.org>
Tested-by: Maria Petrisor <mpetrisor@chromium.org>