diagnostics: Move observer from AddObserver to CreateRoutine
This change removes the method AddObserver on the RoutineControl
interface and moves it as an optional observer to CreateRoutine.
This makes it easier to use the interface properly and not register
observers late in the routine lifetime cycle.

This change also changes all usages of this interface and changes
behaviour where necessary.

BUG=b:291728340
TEST=(cr) FEATURES=test USE="-cros-debug" emerge-brya diagnostics

Change-Id: Ib33a9c50633315dd4f1288b6cc3553e3b2181875
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4701646
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Tested-by: Bastian Kersting <bkersting@google.com>
Commit-Queue: Bastian Kersting <bkersting@google.com>