system_api: Document misleading name in featured.
featured.proto has a |compressed_data| field, which in fact contains the
*decompressed* seed data as currently used.

This field is populated in
components/variations/service/safe_seed_manager.cc, specifically
|SafeSeedManager::GetSafeSeedStateForPlatform|, using
|active_seed_state_->seed_data|. In turn, this is populated from
|SetActiveSeedState|, which comes from |VariationsSeedStore::LoadSeed|
(called by VariationsFieldTrialCreatorBase::CreateTrialsFromSeed).

|VariationsSeedStore::LoadSeed| provides a *decompressed* seed.

It is infeasible to rename the field, as any uprev would immediately
break the build in chromium. The only way to make a rename work would be
to land the change and immediately land a cros_system_api uprev in
chromium that **also** modified all usage of the |compressed_data|
field.  However, DEPS changes should generally not be more complex than
just updating the commit hash (so that if a developer who isn't aware of
the need for the more complex change does the uprev, the build still
won't break).

Alternatively, we could introduce a new field and temporarily accept
both the new and old fields while chromium moves to using the new field.
However, this would also be a complex change.

BUG=None
TEST=CQ

Change-Id: I7919728f7f39d14aac4f9e232af2abd5c9d8ec3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4701767
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Kendrake Tsui <kendraketsui@google.com>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>