cryptohome: Eliminate more DCHECKs
Note: This is mostly a reland of crrev.com/c/4703023 but with unit test
fixes so that the tests should work properly when run with (non-debug)
optimized builds.

In a bunch of these cases I did some quick reading through the code and
I'm convinced that the code would not work correctly (and would either
blow up or do undefined behavior) and so I converted to CHECK.

In some of the cases I decided the check is not really useful (either
it's already handled as an error or actually works correctly) in which
case I got rid of the check. In a few places I got rid of the check and
added proper handling of the error (or proper non-error handing).

The bulk of the remaining DCHECKs are:
* sequence/thread checks, which are probably best left on DCHECK
* path.IsAbsolute() checks in Platform

Note that testing this is a bit tricky. I ran the unit tests of course
but for a lot of them obviously they weren't triggering anyway because
the test would've already been failing. To some extent all we can really
do is watch for an uptick in crashes.

BUG=b:290817945,b:292223540
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I017aad77503714b69068e1024bf9cab76312abc9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4707422
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>