shill: cellular: avoid missing DNS and MTU details from IP settings
The DNS and MTU values reported by ModemManager in the connected
bearer settings are coming from the network via PCOs, so they may be
available regardless of the actual IP addressing method to be used in
the host. The parsing logic is updated to ensure these two values are
processed unconditionally.

BUG=b:292047994
TEST=CQ passes

Change-Id: Ia3400c240d1672fe7b6f891995575285fe782fa7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4702585
Reviewed-by: Madhav <madhavadas@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>