shill: cellular: don't assume IP address is given only for static IP method
When the modem requires the host to run SLAAC itself, shill must
configure the link-local IPv6 address that the network may have
reported before SLAAC is performed. This means that even if the bearer
IP method is reported as kDHCP, for IPv6 we may also have an address
reported.

The logic is updated so that we no longer check the method in use
before attempting to read the IP address. If there is one, we process
it, regardless of the method.

The logic is also updated to optionally receive a gateway IP address,
but no longer mandatory, as it isn't a strict requirement. The gateway
IP address will only be used if there is a modem IP address to use.

BUG=b:292048101
BUG=b:291899554
TEST=CQ passes

Change-Id: I1c112d7aac1d78605d411964406caa717698ea94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4702587
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>