cryptohome: Add mock MountHelperInterface
This is the first step in cleaning up the test that are entangled
between MountHelper and Mount.cc.

Since Mount now happens out of process, this Mock mount is used to
return expected errors, so that we can efficiently test Mount.cc.

There is a duplicate test that was left from previous cleanup, which is
now deleted. The deleted test actually checks functionality in
MountHelper, so it belongs in mount_helper_unittest.cc which is where
the other copy already exists

BUG=b:290897808
TEST=none

Change-Id: Ib605332622a4ebb8ab945d1b379408315eb4c414
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4700806
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>