dlp: Store crtime in the database.
In order to add crtime to the database and primary key,
a new table should be introduced and being used.

The data already stored in the old table should be migrated
when DB is being initialized (if needed) by copying to the
new one and amending with crtime data that is collected by
`EnumerateFiles`.
After that the old table in the database is deleted and
migration is never run again.

BUG=b:290913473
TEST=FEATURES=test emerge-{$BOARD} chromeos-base/dlp

Change-Id: I96bc49d1bfd7c7beef1a29ec41035be795c695db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4700770
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Tested-by: Sergey Poromov <poromov@chromium.org>