lorgnette: Implement explicit scanner handles
In the existing lorgnette API, each method is a self-contained operation
that opens a scanner, sets it up as requested, and closes the scanner.
In order to support the traditional SANE workflow of setting options
one-by-one and retrieving updated configuration, lorgnette needs to
expose an explicit notion of scanner ownership.

To start, add an OpenScanner/CloseScanner pair that creates and destroys
a scanner handle.  The return value from OpenScanner is intended to be
the scanner's current config, but in this CL only an empty configuration
is returned.

To test with real devices, also add a "show_config" verb to
lorgnette_cli that invokes the new functions.

BUG=b:274860707
TEST=lorgnette_cli show_config -scanner="fujitsu:..."

Change-Id: Ibca665e3920d0154244c2f4f6f624de9a65f49a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4738739
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>