lorgnette: Add missing SANE non-empty option types
For options, lorgnete currently supports SANE_Int and truncated versions
of SANE_Fixed.  This adds SANE_Bool and accessors that round-trip
SANE_Fixed through double.

The point of SANE_Fixed is that it can represent an exact set of decimal
values.  In general, converting this to a floating point representation
would introduce a concern of losing precision.  However, on ChromeOS
double is a 64-bit type, so it contains 53 bits of significand.  Since
SANE_Fixed is defined as a 16-bit integer with 16 additional bits of
fraction, a double can fully represent every valid SANE_Fixed value
without rouning errors.  Note that setting a SANE_Fixed option from a
double that doesn't represent an exact SANE_Fixed value will still be
rounded to the nearest valid value, but this is no different from
coercing a double into a fixed representation anyway.

In addition, options that are not active do not contain a meaningful
value.  Since these cannot be tracked within the range of the numeric
types, this also tracks the active status of an option using the
SANE_OPTION_IS_ACTIVE macro.  Inactive options will return a nullopt
instead of a bogus value.

SANE_TYPE_BUTTON and SANE_TYPE_GROUP are not implemented here.  They do
not contain values, so the existing default failures in setters and
getters should cover them implicitly.

BUG=b:274860707
TEST=CQ

Change-Id: I22fc06b1f6bd80f72ae92fd5fbd9d3ceed5b5ec2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4730555
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Liam Huffman <liamhuffman@google.com>