lorgnette: Implement SANE multi-value numeric options
Options of SANE_TYPE_FIXED and SANE_TYPE_INT can contain more than one
value.  We have ignored this in the past because none of the options we
cared about were multi-value.  As lorgnette starts parsing all returned
SANE options, this needs to be prepared to handle these cases.

The most common case is still a single value, so the existing accessor
API is left as-is.  Accessing an option as though it contained a single
number simply returns or sets the value at index 0, which will be the
only value for most options.  Attempting to set multiple values requires
setting or retrieving the whole list with a vector of the correct type.

BUG=b:274860707
TEST=CQ

Cq-Depend: chromium:4730500
Change-Id: I1af7ad3ca4bc92ebbf5a72f4eec3b3a3dda9587c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4730556
Reviewed-by: Liam Huffman <liamhuffman@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>