chromeos-config: Add |use-legacy-custom-label| to RMA config
Add a new field |use-legacy-custom-label| to RMA config. For Shimless
RMA to work with the legacy RMA process, it should know which custom
label tag to write to the device for the legacy process to understand.
Otherwise, when we run Shimless RMA on a device, then run the legacy RMA
process, the toolkit from the factory branch might view the new custom
label tag as an invalid VPD.

BUG=b:292189869
TEST=./regen.sh && ./run_tests.sh

Cq-Depend: chromium:4712785
Change-Id: Ic6f3e13ec28dbc32f704af3b0b360d7fb5cf0025
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4710289
Reviewed-by: Jesse McGuire <jessemcguire@google.com>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>
Tested-by: Cheng-Han Yang <chenghan@chromium.org>