oobe_config: Cleanup metrics file after Rollback
Add StopTrackingRollbackEvents to Rollback cleanup file. Ensure the
metrics file is deleted when Rollback has finished. Events in the
file will be reported as structured metrics before deletion.

The file creation and preservation has not been implemented yet. This
cleanup is on preparation for that implementation to ensure the
deletion of the file.

BUG=b:261850979
TEST=1) FEATURES=test USE="asan fuzzer" emerge-amd64-generic oobe_config
     2) emerge-hatch oobe_config
        Manual test: call rollback_cleanup and ensure the file is
	deleted and the structured metrics are reported as expected
	Manual test: call rollback_cleanup without metrics file works
	as expected.

Change-Id: I2e2f5d28bd4b8e9864084c2279cea1f54716cd66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4710891
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Igor <igorcov@chromium.org>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>