featured: Add comprehensive json parsing testing.
platform-features.json is central to featured's kernel-feature support,
but lacks test coverage.

Add this to make sure that the (somewhat complex) parsing logic does not
break.

Also, fix a parsing bug that the testing uncovered: if
support_check_commands was present, but not a list, we would ignore it
(rather than failing).

BUG=b:291772920
TEST=cros_run_unit_tests --board=jacuzzi \
     --packages=chromeos-base/featured

Change-Id: I3ade1602bd02f3ee73550b01437957b8da6b75e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4711926
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Kendrake Tsui <kendraketsui@google.com>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>