secanomalyd: adds no_new_privs proc percentage UMA
Adds a sandboxing metric for reporting the percentage of processes that
have the no_new_privs bit set, in their status file. The no_new_privs
bit is already included in the ProcEntry attributes, this CL only adds
the code for calculating a percentage and uploading it similar to other
sandboxing metrics.

BUG=b:292579076
TEST=Histogram emits with the correct value.
TEST=All unit tests pass.

Change-Id: Ibbeac6e77badc7a59c761c9ea2a704669069e59c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4711275
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>