cryptohome: Fix init crash in userdataauth fuzzer
This fixes a CHECK-failure that happens in the userdataauth fuzzer
during initialization due to the PlatformFeatures::Initialize not
working.

The code is failing because the mock dbus buses will return null when
constructing object proxies, which the features object doesn't expect.
Normally constructing proxies can't fail; it doesn't matter if dbus
doesn't work or can't connect, that is separate from just creating the
proxy objects.

The fix here is that I set up the mock buses with mock GetProxyObject
instances which will basically do the same thing as the real
GetProxyObject, but with mocks instead. I also added in a call at the
end to destroy the platform features singleton, so that each run is more
hermetic and the first dbus bus doesn't "leak" into subsequent calls.

BUG=b:288950638
TEST=built & ran fuzzer, no CHECK

Change-Id: I9b119151eaca0a754333f8b19340e1cbae52d691
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4718650
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>