patchpanel: ensure ShillClient remembers interface names and indexes
This patch changes the interface index cache inside ShillClient to also
include interface names of shill Devices used for the datapath setup
(iptables, etc). The key is now the "Interface" property of the shill
Device.

For ethernet and WiFi shill Devices there is no behavior change.

For Cellular Devices, this allows to access the name of the primary
multiplexed interface after it has been destroyed when the primary
Network has disconnected. This information is necessary when cleaning up
the datapath setup.

BUG=b:273741099
TEST=FEATURES=test emerge-$BOARD patchpanel

Change-Id: Ie8358a2795481f836b6318c6d951b04eecd9feb6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4721865
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>