multiple: Determine FP login feature at runtime
We use a build flag to guard the under-develop fp login feature now, and
the major mechanisms have been implemented. Now, we can switch to a
runtime logic to determine whether the FP login feature should be
enabled. This isn't going to be launched yet, so a force file that can
be toggled by the developer will be used to enable the feature.

Remove the fp_login use flag.

BUG=b:281958556
TEST=emerge biod

Change-Id: Ib70c83ce1141ae2e1970cec69a1e0bbf29536ee2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4719690
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Cindy Lin <xcl@google.com>
Reviewed-by: Craig Hesling <hesling@chromium.org>