oobe_config: Create CleanStaleRollbackTracking
The Rollback metrics file needs to be deleted if it is stale and the
device is unowned.

Create new method that checks the stale time and calls StopTracking
if 15 days have passed since it was modified. CleanStaleRollbackTracking
will be used in a follow-up CL.

The metrics file is used to keep track of the policy values set for
Rollback and the events that were triggered before powerwash. After
powerwash, it is used to report the old events or read the header. If
Rollback succeeds and the device is re-owned, the Rollback metrics file
is deleted. However, it is possible that after powerwash the device does
not get re-owned.

If there is a reboot when the device is unowned and the Rollback metrics
file has not been modified for 15 days, we need to delete the file to
avoid leaking information.

BUG=b:261850979
TEST=1) FEATURES=test USE="asan fuzzer" emerge-amd64-generic oobe_config
     2) Manual test: trigger code in a device.

Change-Id: I2bae059ac7a79dee7a9f1b02ff60c87806fbfd1e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4720027
Reviewed-by: Igor <igorcov@chromium.org>
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>