patchpanel: Change owner of /run/xtable.lock to patchpaneld
/run/xtable.lock is usually created by iptables to make sure that
only one iptables command is executed at the same time. Currently
iptables commands are used by patchpanel and chromeos-firewall-init [1],
both of them are run as the root user.

This patch uses tmpfile.d to create that lock file and change the owner
of it to patchpaneld, so that iptables can be ran with user patchpaneld.
This is a prerequisite to start patchpanel daemon without root.

[1]
https://crsrc.org/o/src/third_party/chromiumos-overlay/chromeos-base/chromeos-firewall-init/files/iptables.conf

BUG=b:293237848
TEST=CQ dry run;
TEST=Rebooted DUT, observed that /run/xtable.lock is owned by
  patchpaneld.

Cq-Depend: chromium:4719714
Change-Id: I5bc8dfab1498e3e97e8f07333d9ad35c94a20a10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4718273
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>