common-mk: Fix --bin behavior for pre_test
Prior to chromium:4660516, bin used to be empty but unintentionally it
was changed to using test_args[0] which is fine for action=run, but is
not for pre_test where binary is not specified. The original code
wrapped None in an array when test_nin_args = [test_bin] was set,
which became [None] so self.bin=test_args[0] was not error. This is now
fixed when we explicitly set none for binary.

BUG=b:270677405
BUG=b:293327433
TEST=FEATURES=test emerge-zork dev_rust/libchromeos

Change-Id: I4b7d949aad28148549d92c34d5c993844b5512f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4720788
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>
Auto-Submit: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Dennis Kempin <denniskempin@google.com>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>