cryptohome: Handle blank recovery input values
Since promoting some of these DCHECKs to CHECKs we see that these inputs
are not being handled properly. Since these inputs are coming in
externally we should never crash on them. The "doesn't have an input"
CHECKs should be okay because if we don't actually have a recovery input
we shouldn't reach this code, but blanks strings will actually show up.

In the case of the user gaia ID the CHECK is converted to an explicit
error, similar to omitting the username. In the case of the device user
ID we just allow it (and add a unit test to cover this case) since this
appears to be expected to work.

BUG=b:293231020
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Id8229d090357b82ae3bcf9d6db731423a90c13b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4722207
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Anastasiia N <anastasiian@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>