shill: wifi: Only scan once after ensured scan.
This change seeks to simplify the codebase and remove the possibility of
scanning loops that can occur when an ensured scan is queued.

Before this change, the assumption made about the state machine is that
immediately after every scanning state, there will be an idle state.
However, it's valid to go directly from scanning -> connecting.  This
means that it's valid to go from scanning -> connecting -> idle as well
as to go through the transition scanning -> connecting -> connected ->
idle.  This means that the logic that would force a scan when the
connecting or connected states preceeded the idle transition can result
in incorrect scanning loops.

This change removes checking against the previous phy_state, because in
nearly all cases, the interpretation that a connection interrupted the
requested scan either is or can be incorrect, so it's safer to avoid the
possibility of an incorrect scan loop.

BUG=b:293291623
TEST=FEATURES="test emerge-$BOARD shill

Change-Id: I99c01b0301fec343c276cca8657e9f08a5002853
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4721816
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: David Ruth <druth@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: David Ruth <druth@chromium.org>