secanomalyd: refactors process filtering functions
Refactors the process enumeration code such that a list of processes can
be filtered after being produced, and unifies the filtering logic. This
decouples the reading and filtering of the processes, which makes it
possible to filter an existing list of processes without having to
enumerate a new list.
Moves the error for when the init process is not found from the
call-site to the processes library.
Introduces the init_proc_ attribute to the daemon, to avoid having to
read all processes and find it again each time it is needed.
Removes several unused headers.

BUG=b:275758280
TEST=All unit tests pass. Process list in crash report intact.

Change-Id: Ice997e11b73267c797c377a4e71c4d4bb5a578b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4738589
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>