diagnostics: Update the soc_id in cpu_fetcher_test to have a '\n'
When you read this file from sysfs it has a '\n' at the end of
it. Even though the function works both with and without the '\n', we
really want the test to be representative of the real world. Add the
'\n' to the test.

BUG=None
TEST=emerge with FEATURES=test

Change-Id: I11d19409920ca93acbea18204ba82b5a44a57ad5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4727761
Tested-by: Douglas Anderson <dianders@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Kerker Yang <kerker@chromium.org>