diagnostics: Properly show the name of Qualcomm SoCs
The name of Qualcomm SoCs has always been reported slightly
incorrectly by diagnostics. Rather than reporting that devices have a
"SC7180" chip, for instance, the prefix was left off and it was just
reported that we had a "7180" chip. The root of the problem was that
diagnostics was just taking the 4 hexadecimal digits in the JEP106 SoC
ID and assuming that those were the marketing name of the chip. While
historically that was pretty close, it's never been guaranteed.

After firmware updates for sc7180 devices, however, this changed from
slightly wrong to very wrong. The firmware update changed which number
was reported as the JEP106 so that it didn't match any sort of
marketing name for the chip. Now the diagnostics app was reporting
"Qualcomm 01a9" which is very confusing to users.

On Qualcomm SoCs there's actually a separate SoC driver in the kernel
that knows how to provide the marketing name for the chip. Let's
change diagnostics to look for that instead. If, for some reason,
that's not there then we'll still fallback to the old reporting
scheme. Using this new system, sc7180 devices will now report
"Qualcomm Snapdragon SC7180".

Presumably Mediatek devices will also want to do something similar to
this at some point in time.

BUG=b:259402995
TEST=Unit tests; check diagnostics app on lazor device.

Change-Id: Iffbfed5ea2a3e69747b68bf8d45228f3de076080
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4727762
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Ethan Cheng <yycheng@google.com>
Reviewed-by: Kerker Yang <kerker@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>