vm_tools: concierge: timeout aggressive balloon
Add timeout to aggressive balloon. Although the LMKD almost always
aborts the aggressive balloon by trying to kill a high priority process,
sometimes it does not. It is unclear why this happens - one possibility
is that the guest is inactive and fast enough that the balloon never
generates enough PSI to trigger LMKD. If inflating the balloon times
out, assume that that means the balloon is failing to inflate further
(i.e. is fully inflated) and end the aggressive balloon.

BUG=b:259503518
TEST=repeat-test.sh -r 300 -- kohaku arc.Swap
TEST=cros_run_unit_tests --board=$BOARD --packages vm_host_tools

Change-Id: I972f9298ba82ae8650087e2b7aa06cb163cf0a39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4744519
Tested-by: David Stevens <stevensd@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Shin Kawamura <kawasin@google.com>
Reviewed-by: Kameron Lutes <kalutes@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>