diagnostics: Avoid waiting delayed tasks in audio unit tests
The audio_set_volume routine doesn't explicitly set a timeout so we
should not use a delayed task as an indicator of the completion of
routine. Instead, use RunLoop::RunUntilIdle as a workaround.

One might inject a callback to the fake chromium_data_collector.
However, the routine is not yet in the finished state when the data
collector repond (the routine update its status AFTER the response).

BUG=b:293555668
TEST=USE="-cros-debug" FEATURES=test emerge-volteer diagnostics
TEST=Verify the execution time of unit tests reduce from about 6
     seconds to <10 milliseconds.

Change-Id: I2047c5dbae0286a9d741c2cfa0ece8f9e7486ae9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4728651
Reviewed-by: Kerker Yang <kerker@chromium.org>
Tested-by: Wei-Luan Wang <weiluanwang@google.com>
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>