secanomalyd: adds user namespace to ProcEntry
Adds a new ino_t field to ProcEntry to record the user namespace of the
process, and populates it during process enumeration.
The user namespace is parsed from the /proc/pid/ns/user symlink.
This is needed to correctly determine the privilege level of
the processes, and for the forbidden intersection evaluation.
Updates all tests to validate the new ProcEntry attribute.
Updates all tests with more realistic namespace values.
Improves the readability of the constants listed in processes.cc.

BUG=b:293617627
TEST=All tests pass. List of processes in the crash report intact.

Change-Id: Iaf0ac040bb4f2ad258ae332d1cc82bab6e988da6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4729356
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>